refactor(portal): lean shared kernel — move fetch-apps + colors to scenes/common (#2004)
* refactor(portal): move shared fetch-apps query to scenes/common/apps

Both the v2 AppSelector and the v3 AppsDropdown query the team's apps, so
the FetchApps doc moves to scenes/common/apps for a single shared copy.
Pure relocation (git mv, byte-identical) plus import repoints across the 4
consumers and the AppsDropdown test mock. The separate Team/page/Apps
fetch-apps doc is unrelated and untouched.

* refactor(portal): move shared Color types to scenes/common/colors

The Color/ColorName/colors type+const are used by v2 (Header, Profile
ColorSelector, color-atom) and v3 (UserPopup), so they move from
Profile/types.ts to scenes/common/colors. Pure relocation plus import
repoints; calculateColorFromString stays in @/lib and the jotai colorAtom
stays in Portal/layout.
