+++ web/scenes/common/apps/graphql/client/fetch-apps.generated.ts
